requirements.txt
aiodns==1.1.1
aiohttp==2.2.5
aiohttp-jinja2==0.13.0
aiopg==0.13.0
alembic==0.9.5
async-timeout==1.2.1
cchardet==2.1.1
chardet==3.0.4
decorator==4.1.2
ipython==6.1.0
ipython-genutils==0.2.0
jedi==0.10.2
Jinja2==2.9.6
Mako==1.0.7
MarkupSafe==1.0
multidict==3.1.3
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
psycopg2==2.7.3
ptyprocess==0.5.2
pycares==2.3.0
Pygments==2.2.0
python-dateutil==2.6.1
python-editor==1.0.3
PyYAML==3.12
simplegeneric==0.8.1
six==1.10.0
SQLAlchemy==1.1.13
traitlets==4.3.2
wcwidth==0.1.7
yarl==0.12.0
